Allow define USI_TWI_WITHOUT_STATS to exclude stats from compiling
Compiling the stats functions and global variables adds around 300 bytes to
the resulting code. This is quite a blob when the feature is not needed, so
the define USI_TWI_WITHOUT_STATS results in compiling the code without  any
statistics function.

This feature does not bread backwards compatibility, as the define has must
explicitly be stated to change the behaviour.